Eliminating compiler wartnings in 64-bit mode
git-svn-id: https://svn.code.sf.net/p/stlplus/stlplus3@522 ebc0ae9a-4168-4a02-ab27-d0bb227af65f

+++ trunk/persistence/persistent_contexts.cpp
@@ -107,7 +107,7 @@ namespace stlplus
         if (found == m_pointers.end())
         {
           // add a new mapping
-          unsigned magic = m_pointers.size();
+          unsigned magic = static_cast<unsigned>(m_pointers.size());
           m_pointers[pointer] = magic;
           return std::pair<bool,unsigned>(false,magic);
         }
@@ -121,7 +121,7 @@ namespace stlplus
         if (found == m_objects.end())
         {
           // add a new mapping
-          unsigned magic = m_objects.size();
+          unsigned magic = static_cast<unsigned>(m_objects.size());
           m_objects[pointer] = magic;
           return std::pair<bool,unsigned>(false,magic);
         }
